requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.7.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
